+++ run/download_data.py
@@ -40,9 +40,9 @@ assert sys.version_info.major >= 3, \
 "ERROR: Python 3 is required to run download_data.py!"
 
 # Define global variables
-GEOSCHEM_INPUT_FILE = "./geoschem_config.yml"
 DATA_DOWNLOAD_SCRIPT = "./auto_generated_download_script.sh"
-
+CONFIG_FILE = "./HEMCO_sa_Config.rc"
+TIME_FILE = "./HEMCO_sa_Time.rc"
 
 def read_config_file(
         config_file,
@@ -115,9 +115,6 @@ def extract_pathnames_from_log(
 
         while line:
 
-            # Replace double slashes with single slash
-            line = line.replace("CHEM_INPUTS//", "CHEM_INPUTS/")
-
             # Convert line to uppercase for string match
             upcaseline = line.upper()
 
@@ -125,12 +122,12 @@ def extract_pathnames_from_log(
             if (": OPENING" in upcaseline) or (": READING" in upcaseline):
                 data_found.add(line.split()[-1])
 
-                # Search for data paths that are missing
+            # Search for data paths that are missing
             elif "FILE NOT FOUND" in upcaseline:
                 data_missing.add(line.split()[-1])
 
-                # Search for certain dry-run comment strings
-                # (and make sure to prevent duplicates)
+            # Search for certain dry-run comment strings
+            # (and make sure to prevent duplicates)
             elif ("!!! STA" in upcaseline) or ("!!! END" in upcaseline) or \
                  ("!!! SIM" in upcaseline) or ("!!! MET" in upcaseline) or \
                  ("!!! GRI" in upcaseline):
@@ -156,6 +153,8 @@ def extract_pathnames_from_log(
             if "ExtData" in path:
                 index = path.find("ExtData")
                 local_prefix = path[:index]
+                if "ExtData" not in local_prefix:
+                    local_prefix = os.path.join(local_prefix, "ExtData")
                 break
 
         # Exit if the local path does not contain ExtData
@@ -189,140 +188,35 @@ def get_run_info():
         start_time, end_date, end_time, met, grid, and sim.
     """
 
-    # Read GEOS-Chem configuration file
-    config = read_config_file(
-        GEOSCHEM_INPUT_FILE,
-        to_str=True
-    )
-
     # Create dictionary with GEOS-Chem simulation parameters
     # NOTE: Numbers are returned as strings, and need to be converted
     run_info = {}
-    run_info["nest"] = ""
-    run_info["tomas15"] = False
-    run_info["tomas40"] = False
-    run_info["start_date"] = int(
-        config["simulation"]["start_date"][0]
-    )
-    run_info["start_time"] = int(
-        config["simulation"]["start_date"][1]
-    )
-    run_info["end_date"] = int(
-        config["simulation"]["end_date"][0]
-    )
-    run_info["end_time"] = int(
-        config["simulation"]["end_date"][1]
-    )
-    run_info["met_field"] = config["simulation"]["met_field"]
-    run_info["sim"] = config["simulation"]["name"]
-    run_info["resolution"] = config["grid"]["resolution"]
-    run_info["grid"] = get_grid_suffix(
-        run_info["resolution"]
-    )
-    run_info["nest"] = get_nest_suffix(
-        config["grid"]["longitude"]["range"]
-    )
-    run_info["tomas15"] = \
-        "NK15" in config["operations"]["transport"]["transported_species"]
-    run_info["tomas40"] = \
-        "NK40" in config["operations"]["transport"]["transported_species"]
+
+    # Read start & end datetimes
+    config = read_config_file(TIME_FILE, to_str=True)
+    temp = config["START"].replace("-","").replace(":","").split()
+    run_info["start_date"] = int(temp[0])
+    run_info["start_time"] = int(temp[1])
+    temp = config["END"].replace("-","").replace(":","").split()
+    run_info["end_date"] = int(temp[0])
+    run_info["end_time"] = int(temp[1])
+
+    # Get root data dir, met field, and resolution
+    with open(CONFIG_FILE, "r", encoding="UTF-8") as ifile:
+        for line in ifile:
+            line = line.strip()
+            if "END SECTION SETTINGS" in line:
+                break
+            if "ROOT:" in line:
+                run_info["root_data_dir"] = line.split(":")[-1].strip(" ")
+            if "MET:" in line:
+                run_info["met_field"] = line.split(":")[-1].strip(" ")
+            if "RES:" in line:
+                run_info["resolution"] = line.split(":")[-1].strip(" ")
+                run_info["grid"] = run_info["resolution"]
 
     return run_info
 
-
-def get_grid_suffix(
-        resolution
-):
-    """
-    Given a model resolution, returns the grid filename suffix.
-
-    Args:
-    -----
-    resolution : str
-        The grid resolution (read from geoschem_config.yml)
-
-    Returns:
-    --------
-    suffix : str
-        The corresponding filename suffix
-    """
-    if "4.0x5.0" in resolution:
-        return "4x5"
-    if "2.0x2.5" in resolution:
-        return "2x25"
-    if "0.5x0.625" in resolution:
-        return "05x0625"
-    return "025x03125"
-
-
-def get_nest_suffix(
-        longitude
-):
-    """
-    Given a model resolution, returns the nested-grid suffix.
-
-    Args:
-    -----
-    resolution : str
-        The grid resolution (read from geoschem_config.yml)
-
-    Returns:
-    --------
-    suffix : str
-        The corresponding nested-grid suffix
-    """
-    if "-130" in longitude or "-140" in longitude:
-        return "na"
-    if "60" in longitude or "70" in longitude:
-        return "as"
-    return ""
-
-
-def get_remote_restart_filename(
-        local_prefix,
-        run_info,
-        rst_info
-):
-    """
-    Returns the remote restart file name for a given
-    GEOS-Chem Classic simulation type.
-
-    Args:
-    -----
-    local_prefix : str
-        The root data folder.  ExtData is a subfolder of this folder.
-    run_info : dict
-        Information read from geoschem_config.yml
-    rst_info : dict
-        Restart file paths (local and remote), as read from the
-        download_data.yml configuration file.
-
-    Returns:
-    --------
-    remote_rst : str
-        Path to the remote restart file
-    """
-
-    # Simulation type
-    simulation = run_info["sim"].lower()
-
-    # Remote restart file directory
-    root = os.path.join(local_prefix, "ExtData", rst_info["root"])
-
-    # Special handling for fullchem
-    if "fullchem" in simulation:
-        if run_info["tomas15"] is True:
-            return os.path.join(root, rst_info["tomas15"]["remote"])
-        if run_info["tomas40"] is True:
-            return os.path.join(root, rst_info["tomas40"]["remote"])
-        return os.path.join(root, rst_info["fullchem"]["remote"])
-
-    # Special handling for mercury
-    if "mercury" in simulation or "hg" in simulation:
-        return os.path.join(root, rst_info["mercury"]["remote"])
-
-    # All other simulations use the lowercase simulation name
-    return os.path.join(root, rst_info[simulation]["remote"])
 
 
 def replace_entry_in_list(
@@ -348,66 +242,6 @@ def replace_entry_in_list(
         The modified list
     """
     return list(map(lambda x: x.replace(old_entry, new_entry), the_list))
-
-
-def expand_restart_file_names(
-        paths,
-        args,
-        run_info
-):
-    """
-    Tests if the GEOS-Chem restart file is a symbolic link to
-    ExtData.  If so, will append the link to the remote file
-    to the line in which the restart file name is found.
-
-    Args:
-    ----
-    paths : dict
-        Contains output from function extract_pathnames_from_log.
-    args : dict
-        Contains output from function parse_args.
-    run_info : dict
-        Contains output from function get_run_info.
-    """
-
-    # Get the name of the remote restart file for this simulation
-    remote_rst = get_remote_restart_filename(
-        paths["local_prefix"],
-        run_info,
-        args["config"]["restarts"]
-    )
-
-    # First, look for the restart file name in the found files
-    do_exit = False
-    for path in paths["found"]:
-        if "GEOSChem.Restart" in path:
-            new_path = path + " --> " + remote_rst
-            do_exit = True
-            paths["found"] = replace_entry_in_list(
-                paths["found"],
-                path,
-                new_path
-            )
-            break
-    paths["found"] = sorted(paths["found"])
-    if do_exit:
-        return paths
-
-    # Then, look for the restart file name in the missing files
-    for path in paths["missing"]:
-        if "GEOSChem.Restart" in path:
-            new_path = path + " --> " + remote_rst
-            do_exit = True
-            paths["missing"] = replace_entry_in_list(
-                paths["missing"],
-                path,
-                new_path
-            )
-            break
-    paths["missing"] = sorted(paths["missing"])
-
-    # Return the updated data paths
-    return paths
 
 
 def write_unique_paths(
@@ -479,170 +313,23 @@ def create_download_script(
         # Write download commands for only the missing data files
         for path in paths["missing"]:
 
-            if "-->" in path:
+            # We do not download HEMCO restart files
+            if "HEMCO_restart" in path:
+                continue
 
-                # ------------------------------------------------------
-                # Edge case: Linked restart files
-                # ------------------------------------------------------
-
-                # First copy the restart file to local ExtData
-                remote_rst = (path.split("-->")[1]).strip()
-                local_rst = (path.split("-->")[0]).strip()
-                index2 = remote_rst.find("ExtData") + 7
-                prefix = local_rst
-                extdata = remote_rst[:index2]
-                remote_rst = remote_root + remote_rst[index2:]
-                cmd = cmd_prefix + quote + remote_rst + quote
-                if is_s3_bucket:
-                    cmd += " " + prefix
-                print(cmd, file=ofile)
-                print(file=ofile)
-
-                # If the file does not exist in the run directory,
-                # then copy it from the restart folder.
-                # This only has to be done if not using the amazon portal.
-                if not is_s3_bucket:
-                    if not os.path.exists(local_rst):
-                        index3 = remote_rst.find("GEOSCHEM_RESTARTS")
-                        rst = os.path.join(extdata, remote_rst[index3:])
-                        cmd = "cp -f " + rst + " " + local_rst
-                        print(cmd, file=ofile)
-                        print(file=ofile)
-
-            elif "gmi.clim.IPMN.geos5.2x25.nc" in path:
-
-                # ------------------------------------------------------
-                # Edge case: GMI IPMN file is really the PMN file
-                # ------------------------------------------------------
-
-                # Download the PMN file
-                index = path.find("ExtData") + 7
-                local_dir = os.path.dirname(path)
-                remote_path = remote_root + path[index:]
-                remote_path = remote_path.replace("IPMN", "PMN")
-                cmd = cmd_prefix + quote + remote_path + quote
-                if is_s3_bucket:
-                    cmd += " " + local_dir + "/"
-                print(cmd, file=ofile)
-
-                # Rename it to IPMN
-                cmd = "mv " + local_dir + "/gmi.clim.PMN.geos5.2x25.nc " + \
-                      local_dir + "/gmi.clim.IPMN.geos5.2x25.nc"
-                print(cmd, file=ofile)
-
-            elif "gmi.clim.NPMN.geos5.2x25.nc" in path:
-
-                # ------------------------------------------------------
-                # Edge case: GMI NPMN file is really the PMN file
-                # ------------------------------------------------------
-
-                # Download the PMN file
-                index = path.find("ExtData") + 7
-                local_dir = os.path.dirname(path)
-                remote_path = remote_root + path[index:]
-                remote_path = remote_path.replace("NPMN", "PMN")
-                cmd = cmd_prefix + quote + remote_path + quote
-                if is_s3_bucket:
-                    cmd += " " + local_dir + "/"
-                print(cmd, file=ofile)
-
-                # Rename it to NPMN
-                cmd = "mv " + local_dir + "/gmi.clim.PMN.geos5.2x25.nc " + \
-                      local_dir + "/gmi.clim.NPMN.geos5.2x25.nc"
-                print(cmd, file=ofile)
-                print(file=ofile)
-
-            elif "gmi.clim.RIPA.geos5.2x25.nc" in path:
-
-                # ------------------------------------------------------
-                # Edge case: GMI RIPA file is really the RIP file
-                # ------------------------------------------------------
-
-                # Download the RIP file
-                index = path.find("ExtData")+7
-                local_dir = os.path.dirname(path)
-                remote_path = remote_root + path[index:]
-                remote_path = remote_path.replace("RIPA", "RIP")
-                cmd = cmd_prefix + quote + remote_path + quote
-                if is_s3_bucket:
-                    cmd += " " + local_dir + "/"
-                print(cmd, file=ofile)
-
-                # Rename it to NPMN
-                cmd = "mv " + local_dir + "/gmi.clim.RIP.geos5.2x25.nc " + \
-                      local_dir + "/gmi.clim.RIPA.geos5.2x25.nc"
-                print(cmd, file=ofile)
-                print(file=ofile)
-
-            elif "gmi.clim.RIPB.geos5.2x25.nc" in path:
-
-                # ------------------------------------------------------
-                # Edge case: GMI RIPB file is really the RIP file
-                # ------------------------------------------------------
-
-                # Download the RIP file
-                index = path.find("ExtData")+7
-                local_dir = os.path.dirname(path)
-                remote_path = remote_root + path[index:]
-                remote_path = remote_path.replace("RIPB", "RIP")
-                cmd = cmd_prefix + quote + remote_path + quote
-                if is_s3_bucket:
-                    cmd += " " + local_dir + "/"
-                print(cmd, file=ofile)
-
-                # Rename it to RIPB
-                cmd = "mv " + local_dir + "/gmi.clim.RIP.geos5.2x25.nc " + \
-                      local_dir + "/gmi.clim.RIPB.geos5.2x25.nc"
-                print(cmd, file=ofile)
-                print(file=ofile)
-
-            elif "gmi.clim.RIPD.geos5.2x25.nc" in path:
-
-                # ------------------------------------------------------
-                # Edge case: GMI RIPD file is really the RIP file
-                # ------------------------------------------------------
-
-                # Download the RIP file
-                index = path.find("ExtData")+7
-                local_dir = os.path.dirname(path)
-                remote_path = remote_root + path[index:]
-                remote_path = remote_path.replace("RIPD", "RIP")
-                cmd = cmd_prefix + quote + remote_path + quote
-                if is_s3_bucket:
-                    cmd += " " + local_dir + "/"
-                print(cmd, file=ofile)
-
-                # Rename it to RIPD
-                cmd = "mv " + local_dir + "/gmi.clim.RIP.geos5.2x25.nc " + \
-                      local_dir + "/gmi.clim.RIPD.geos5.2x25.nc"
-                print(cmd, file=ofile)
-                print(file=ofile)
-
-            elif "ExtData" in path:
-
-                # ------------------------------------------------------
-                # All other files in ExtData
-                # ------------------------------------------------------
-                index = path.find("ExtData") + 7
-                local_dir = os.path.dirname(path)
-                remote_path = remote_root + path[index:]
-                cmd = cmd_prefix + quote + remote_path + quote
-                if is_s3_bucket:
-                    cmd += " " + local_dir + "/"
-                print(cmd, file=ofile)
-                print(file=ofile)
-
-        # Kludge: Create a ExtData/CHEM_INPUTS folder if it
-        # does not exist. This will prevent abnormal exits.
-        chem_dir = paths["local_prefix"] + 'ExtData/CHEM_INPUTS'
-        cmd = f"if [[ ! -d {chem_dir} ]]; then mkdir {chem_dir}; fi"
-        print(cmd, file=ofile)
-        print(file=ofile)
+            # Write path
+            index = path.find("ExtData") + 7
+            local_dir = os.path.dirname(path)
+            remote_path = remote_root + path[index:]
+            cmd = cmd_prefix + quote + remote_path + quote
+            if is_s3_bucket:
+                cmd += " " + local_dir + "/"
+            print(cmd, file=ofile)
+            print(file=ofile)
 
         # Close file and make it executable
         ofile.close()
         os.chmod(DATA_DOWNLOAD_SCRIPT, 0o755)
-
 
 def download_the_data(
         args
@@ -663,7 +350,6 @@ def download_the_data(
     # Get a unique list of data paths, both found and missing:
     # Expand the data paths to include links to restart files
     paths = extract_pathnames_from_log(args)
-    paths = expand_restart_file_names(paths, args, run_info)
 
     # Write a list of unique file paths
     write_unique_paths(paths, args["dryrun_log"] + ".unique")